Gracia Final support 5/6
git-svn-id: http://svn.l2jserver.com/trunk@2949 d34dfa0f-f91c-0410-86dd-cf06f19dd0ad

+++ L2_GameServer/java/net/sf/l2j/gameserver/network/serverpackets/ExShowCropInfo.java
@@ -71,9 +71,9 @@ public class ExShowCropInfo extends L2GameServerPacket
 		for (CropProcure crop : _crops)
         {
 			writeD(crop.getId());          // Crop id
-			writeD(crop.getAmount());      // Buy residual
-			writeD(crop.getStartAmount()); // Buy
-			writeD(crop.getPrice());       // Buy price
+			writeD((int)crop.getAmount());      // Buy residual
+			writeD((int)crop.getStartAmount()); // Buy
+			writeD((int)crop.getPrice());       // Buy price
 			writeC(crop.getReward());      // Reward
 			writeD(L2Manor.getInstance().getSeedLevelByCrop(crop.getId())); // Seed Level
 			writeC(1); // rewrad 1 Type
