+++ L2_GameServer/java/net/sf/l2j/gameserver/network/serverpackets/ExShowSeedInfo.java
@@ -68,15 +68,15 @@ public class ExShowSeedInfo extends L2GameServerPacket {
         writeD(_seeds.size());
         for (SeedProduction seed : _seeds)
         {
-            writeD(seed.getId()); // Seed id
-            writeD(seed.getCanProduce()); // Left to buy
-            writeD(seed.getStartProduce()); // Started amount
-            writeD(seed.getPrice());        // Sell Price
-            writeD(L2Manor.getInstance().getSeedLevel(seed.getId())); // Seed Level
-            writeC(1); // reward 1 Type
-            writeD(L2Manor.getInstance().getRewardItemBySeed(seed.getId(),1)); // Reward 1 Type Item Id
-            writeC(1); // reward 2 Type
-            writeD(L2Manor.getInstance().getRewardItemBySeed(seed.getId(),2)); // Reward 2 Type Item Id
+        	 writeD(seed.getId()); // Seed id
+             writeD((int)seed.getCanProduce()); // Left to buy
+             writeD((int)seed.getStartProduce()); // Started amount
+             writeD((int)seed.getPrice());        // Sell Price
+             writeD(L2Manor.getInstance().getSeedLevel(seed.getId())); // Seed Level
+             writeC(1); // reward 1 Type
+             writeD(L2Manor.getInstance().getRewardItemBySeed(seed.getId(),1)); // Reward 1 Type Item Id
+             writeC(1); // reward 2 Type
+             writeD(L2Manor.getInstance().getRewardItemBySeed(seed.getId(),2)); // Reward 2 Type Item Id
         }
     }
